feat: Initialize project with very_good_cli
This commit initializes a new Flutter project using very_good_cli, a tool for building high-quality Flutter applications. The project structure follows best practices and includes a sample counter app, as well as a basic set of testing and analysis tools.

The following packages were added to the project:
- very_good_cli
- flutter_test
-flutter_bloc: ^8.1.2
- flutter_native_splash
-very_good_analysis: ^4.0.0

The project is ready to be built and run on both iOS and Android devices.

Closes #1